Add files via upload
This is an AI based talent scouting system developed on Google Colab using python and machine learning. Users will be able to upload video in the form of .mp4 on their google drive and then the video will be analyzed properly. After that, the AI model will give feedback to the user based on data analysis and so on. It includes more features that will help the user to improve upon.